Ejercicio 2, Reto 15

+++ Repaso/src/com/cmc/repaso/entidades/Validacion.java
@@ -1,7 +1,7 @@
 package com.cmc.repaso.entidades;
 
 public class Validacion {
-	public boolean validarMonto (int monto) {
+	public boolean validarMonto (double monto) {
 		if (monto > 0) {
 			return true;
 		}else {
